Minor changes to flash tree by wa1tnr
Wanted 0x10040000  base address - seems to work.

Writes to flashROM at that address.

Since the program only loads into SRAM, your
resident program is unaffected, and can be
used (later) to inspect what was written to
the flash.  CamelForth was used for this,
locally (see wa1tnr repos on github).

Included hi-flash.uf2 equivalent to your own
compiled version of this program.

	new file:   hi-flash/Makefile
	new file:   hi-flash/README.md
	new file:   hi-flash/hi-flash.uf2
	new file:   hi-flash/inc/flash.h
	new file:   hi-flash/inc/gpio.h
	new file:   hi-flash/inc/resets.h
	new file:   hi-flash/inc/rp2040.h
	new file:   hi-flash/inc/uart.h
	new file:   hi-flash/memory.ld
	new file:   hi-flash/mkuf2
	new file:   hi-flash/src/flash.c
	new file:   hi-flash/src/gpio.c
	new file:   hi-flash/src/init.c
	new file:   hi-flash/src/resets.c
	new file:   hi-flash/src/start.s
	new file:   hi-flash/src/uart.c

On branch dvlp-a